Final pushes for the spring semester. Code cleanup

+++ Application-Code/app/src/main/java/com/example/medication_reminder_android_app/UserInputHandler/AppointmentInputHandler.java
@@ -62,7 +62,7 @@ class AppointmentInputHandler extends InputHandler {
     }
 
     @Override
-    void acknowledgeNotificationRequest(int reminderID, boolean dismissed) {
+    void acknowledgeNotificationRequest(long reminderID, boolean dismissed) {
 
     }
